split query-engine::execution into query and mutation mods
