+++ crates/query-engine/execution/src/lib.rs
@@ -2,6 +2,7 @@
 //! See `/architecture.md#execution` in the repository for more details.
 
 pub mod error;
-pub mod execution;
 pub mod helpers;
 pub mod metrics;
+pub mod mutation;
+pub mod query;
